server/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg[binary]==3.2.9
django-cors-headers==4.3.1
Pillow==11.3.0
python-decouple==3.8
django-filter==23.5
requests==2.31.0
drf-yasg==1.21.7 